feat(mention): Added mention insert support | Minor refactor
