+++ src/App.tsx
@@ -1,7 +1,6 @@
 import { useState } from 'react';
 import React from 'react';
 import DraftEditor from './DraftEditor';
-import CustomMentionEditor from './Mention';
 
 export const App = () => {
     const [content, setContent] = useState(
