--- src/MentionData.ts
@@ -1,65 +1,0 @@
-import { MentionData } from '@draft-js-plugins/mention';
-
-const mentionDataImg: MentionData[] = [
-  {
-    name: 'Matthew Russell',
-    title: 'Senior Software Engineer',
-    avatar:
-      'https://pbs.twimg.com/profile_images/517863945/mattsailing_400x400.jpg',
-  },
-  {
-    name: 'Julian Krispel-Samsel',
-    title: 'United Kingdom',
-    avatar: 'https://avatars2.githubusercontent.com/u/1188186?v=3&s=400',
-  },
-  {
-    name: 'Jyoti Puri',
-    title: 'New Delhi, India',
-    avatar: 'https://avatars0.githubusercontent.com/u/2182307?v=3&s=400',
-  },
-  {
-    name: 'Max Stoiber',
-    title:
-      'Travels around the world, brews coffee, skis mountains and makes stuff on the web.',
-    avatar: 'https://avatars0.githubusercontent.com/u/7525670?s=200&v=4',
-  },
-  {
-    name: 'Nik Graf',
-    title: 'Passionate about Software Architecture, UX, Skiing & Triathlons',
-    avatar: 'https://avatars0.githubusercontent.com/u/223045?v=3&s=400',
-  },
-  {
-    name: 'Pascal Brandt',
-    title: 'HeathIT hacker and researcher',
-    avatar:
-      'https://pbs.twimg.com/profile_images/688487813025640448/E6O6I011_400x400.png',
-  },
-  {
-    name: 'Łukasz Bąk',
-    title: 'Randomly Generated User',
-    avatar: 'https://randomuser.me/api/portraits/men/36.jpg',
-  },
-];
-
-const mentionDataText: MentionData[] = [
-  {
-    name: '#{{Matthew Russell}}',
-    label: 'Matthew Russell',
-    title: 'Senior Software Engineer',
-    subTitle:"India"
-  },
-  {
-    name: '#{{Julian Krispel-Samsel}}',
-    label: 'Julian Krispel-Samsel',
-    title: 'United Kingdom',
-    subTitle:"US"
-  },
-  {
-    name: '#{{Jyoti Puri}}',
-    label: 'Jyoti Puri',
-    title: 'New Delhi, India',
-    subTitle:"UK"
-  },
-];
-
-export const MentionDataValue= {mentionDataImg, mentionDataText};
